Don't attempt to output containers when using 'rm'
https://github.com/buildkite-plugins/docker-compose-buildkite-plugin/pull/211
added a '--rm' to the docker-compose commands in addition to the
container cleanup done in the pre-exit hook.

This appears to silently break the check for linked log file uploads as
no containers will be returned by `docker ps`, and conflicts with
the changes in https://github.com/buildkite-plugins/docker-compose-buildkite-plugin/pull/212
to output failed containers for the same reason, though the later fails
loudly as no containers are passed to `docker inspect` (the former uses
a more graceful loop ).

The test suite was also broken due to conflicts between these two
changes.

To fix things this tests for the 'rm' configuration and doesn't attempt
to output failed containers or upload logs when present. That's not
ideal as it's a bit unexpected; it would probably be worth reconsidering
whether 'rm' should be enabled by default, implementing a more
graceful version of the failed container output check, and tidying up
all the nested and branching conditionals here.

See https://github.com/buildkite-plugins/docker-compose-buildkite-plugin/issues/220