Check service names in cache-from against the services being built
The service names in the cache-from lines are used to place
`cache_from: ...` configurations into the appropriate place in the
modified docker-compose config. Without that `cache_from`, the pulled
image cannot be used for caching, and so all the effort there is
useless.

Previously, a configuration like:

    build:
    - foo
    cache_from:
    - foo-bar-baz:example.com/foo-bar-baz:latest

would download the image, but be unable to use it for caching
purposes. This patch makes the plugin flag this as a soft error,
alerting the user that the `foo-bar-baz` cache-from line doesn't
correspond to any of the services being built.