Add use-prior-image to completely skip repetitive builds
This change adds a new plugin build option, `use-prior-image`. When set
to `true`, we'll check to see if a plugin with the expected image name
(the name specified by the cache-from parameter) exists on the
repository with `docker manifest inspect`. If it exists, we'll tag the
build metadata with the image name for the service and skip any further
pulling or building activity.

For builds where dependencies are built in a service, this can allow for
significant time shaved off of a blocking early build step.