Allow reuse and complete skip of build if prior image exists
This change allows a properly tagged image already pushed to the
registry to tell the plugin to skip any further downloading or building
operations. The plugin will mark the existing tag with build metadata
and finish early. This step can save up to a minute of processing when
we have dependency or asset images that we've already built and pushed
in prior builds.