Fix argument passing
The builder.docker-opts and builder.platform arguments didn't work.
The switch and its values were added to the builder_instance_args array
as a single element, but the switch should be one element and its values
the another. Otherwise it produces an invalid parameter when passed to
the docker commands and errors are thrown.