Add `builder.use-once` parameter
When `builder.use` is `true`, this plugin calls `docker buildx use` which sets the default builder for the entire Buildkite agent.

This can cause issues for other Buildkite steps on the same agent. For example, another pipeline step might land on the same agent and expect to use the default `docker` driver, which does not require the user to specify `--load` when calling `docker build`. This results in the image not being available for subsequent commands (e.g. `docker tag` or `docker push`) and can break a pipeline that was previously working.

This change adds the `builder.use-once` parameter. When `true`, the plugin will pass the `--builder` flag to `docker build` instead of calling `docker buildx use`. This ensures that the specified builder is only used for this particular step.